refactor(core): delete vestigial rendering forwarders from page.computed
`compute_excerpt` and `compute_meta_description` in
`bengal/core/page/computed.py` were deprecated re-exports that forwarded
to `bengal.rendering.page_content` with no added logic. The live
`Page.excerpt` / `Page.meta_description` properties already call the
rendering helpers directly (`runtime.py:669,681`), so the wrappers had
zero production callers — their only importer was a test module.

Per the greenfield-design test (CLAUDE.md / epic-delete-forwarding-wrappers),
these are vestigial forwarders that should have been deleted when the
`page_content` extraction happened. Removing them keeps rendering-derived
behavior out of core.

- Delete both functions and their docstring entries from page.computed.
- Move their behavioral tests (which carried unique coverage: max-160-chars,
  mid-markdown truncation, headers-in-content) from
  tests/unit/core/test_computed_functions.py to
  tests/unit/rendering/test_page_content.py, importing the canonical
  functions from bengal.rendering.page_content.

No production migration needed. Full unit suite green (2997 passed).

Found via a core-boundary-audit dynamic workflow; both findings confirmed
3/3 by independent adversarial verifiers and re-verified by hand.
