+++ tests/unit/core/test_computed_functions.py
@@ -3,7 +3,7 @@ Tests for computed free functions in bengal.core.page.computed.
 
 These test the functions directly (not through Page property wrappers),
 verifying correctness of age calculations, author parsing, series parsing,
-series navigation, and the legacy content-derived rendering wrappers.
+and series navigation.
 
 Completes Track 4b of the free-threading migration epic.
 """
@@ -17,8 +17,6 @@ from typing import Any
 from bengal.core.page.computed import (
     compute_age_days,
     compute_age_months,
-    compute_excerpt,
-    compute_meta_description,
     compute_reading_time,
     compute_word_count,
     get_all_authors,
@@ -49,34 +47,6 @@ class TestComputeWordCount:
 
 
 # -----------------------------------------------------------------------
-# compute_meta_description
-# -----------------------------------------------------------------------
-
-
-class TestComputeMetaDescription:
-    """Direct tests for the compute_meta_description compatibility wrapper."""
-
-    def test_explicit_description_preferred(self):
-        result = compute_meta_description(
-            {"description": "Explicit desc"},
-            "Some long content",
-        )
-        assert result == "Explicit desc"
-
-    def test_generated_from_content(self):
-        result = compute_meta_description({}, "Short content here.")
-        assert result == "Short content here."
-
-    def test_empty_content_no_description(self):
-        assert compute_meta_description({}, "") == ""
-
-    def test_max_160_chars(self):
-        long = "This is a sentence. " * 30
-        result = compute_meta_description({}, long)
-        assert len(result) <= 160
-
-
-# -----------------------------------------------------------------------
 # compute_reading_time
 # -----------------------------------------------------------------------
 
@@ -94,63 +64,6 @@ class TestComputeReadingTime:
     def test_rounds(self):
         assert compute_reading_time(250) == 1  # 1.25 -> 1
         assert compute_reading_time(350) == 2  # 1.75 -> 2
-
-
-# -----------------------------------------------------------------------
-# compute_excerpt
-# -----------------------------------------------------------------------
-
-
-class TestComputeExcerpt:
-    """Direct tests for the compute_excerpt compatibility wrapper.
-
-    The rendering-side helper returns rendered HTML, strips leading h1, and
-    truncates at ~250 chars before rendering. The core wrapper preserves the
-    older import path without owning that rendering behavior.
-    """
-
-    def test_short_content_renders_markdown(self):
-        result = compute_excerpt("Hello world.")
-        assert "Hello world" in result
-        assert "<" in result  # Rendered to HTML (e.g. <p>)
-
-    def test_empty_string(self):
-        assert compute_excerpt("") == ""
-
-    def test_strips_leading_h1(self):
-        result = compute_excerpt("# My Title\n\nFirst paragraph here.")
-        assert "First paragraph" in result
-        assert "My Title" not in result
-
-    def test_renders_markdown(self):
-        result = compute_excerpt("This has **bold** and *italic*.")
-        assert "bold" in result
-        assert "<strong>" in result or "**" not in result  # Rendered
-
-    def test_does_not_cut_mid_markdown(self):
-        """Excerpt should not end with orphaned ** or other markdown syntax."""
-        content = (
-            "Bengal is a powerful static site generator. "
-            "Key Features: Fast Builds, Asset Optimization, **SEO** friendly."
-        )
-        result = compute_excerpt(content)
-        from bengal.core.utils.text import strip_html
-
-        plain = strip_html(result)
-        assert not plain.endswith("**"), "Should not end with orphaned **"
-
-    def test_headers_in_content(self):
-        """Headers in content are rendered and available for card excerpt."""
-        content = (
-            "Intro paragraph here.\n\n"
-            "## Key Features\n\n"
-            "### Fast Builds\n"
-            "- Parallel processing\n"
-            "- Asset Optimization\n"
-            "**SEO** friendly."
-        )
-        result = compute_excerpt(content)
-        assert "Key Features" in result or "Intro" in result
 
 
 # -----------------------------------------------------------------------
