+++ tests/unit/rendering/test_page_content.py
@@ -87,10 +87,6 @@ def test_excerpt_fallback_renders_markdown_without_leading_h1() -> None:
     assert "<strong>" in excerpt
 
 
-def test_compute_excerpt_handles_empty_source() -> None:
-    assert compute_excerpt("") == ""
-
-
 def test_ast_cache_helpers_handle_document_dict() -> None:
     ast_cache = {
         "type": "document",
@@ -111,3 +107,87 @@ def test_page_private_content_shims_delegate_to_rendering_helpers() -> None:
     assert page._extract_links_from_ast() == ["/guide/"]
     assert page._extract_text_from_ast() == ""
     assert page._strip_html_to_text("<p>Hello</p>") == strip_html_to_text("<p>Hello</p>")
+
+
+# -----------------------------------------------------------------------
+# compute_meta_description
+# -----------------------------------------------------------------------
+
+
+class TestComputeMetaDescription:
+    """Direct tests for compute_meta_description (raw-content path)."""
+
+    def test_explicit_description_preferred(self):
+        result = compute_meta_description(
+            {"description": "Explicit desc"},
+            "Some long content",
+        )
+        assert result == "Explicit desc"
+
+    def test_generated_from_content(self):
+        result = compute_meta_description({}, "Short content here.")
+        assert result == "Short content here."
+
+    def test_empty_content_no_description(self):
+        assert compute_meta_description({}, "") == ""
+
+    def test_max_160_chars(self):
+        long = "This is a sentence. " * 30
+        result = compute_meta_description({}, long)
+        assert len(result) <= 160
+
+
+# -----------------------------------------------------------------------
+# compute_excerpt
+# -----------------------------------------------------------------------
+
+
+class TestComputeExcerpt:
+    """Direct tests for compute_excerpt (raw-content path).
+
+    The helper returns rendered HTML, strips a leading h1, and truncates at
+    ~250 chars before rendering.
+    """
+
+    def test_short_content_renders_markdown(self):
+        result = compute_excerpt("Hello world.")
+        assert "Hello world" in result
+        assert "<" in result  # Rendered to HTML (e.g. <p>)
+
+    def test_empty_string(self):
+        assert compute_excerpt("") == ""
+
+    def test_strips_leading_h1(self):
+        result = compute_excerpt("# My Title\n\nFirst paragraph here.")
+        assert "First paragraph" in result
+        assert "My Title" not in result
+
+    def test_renders_markdown(self):
+        result = compute_excerpt("This has **bold** and *italic*.")
+        assert "bold" in result
+        assert "<strong>" in result or "**" not in result  # Rendered
+
+    def test_does_not_cut_mid_markdown(self):
+        """Excerpt should not end with orphaned ** or other markdown syntax."""
+        content = (
+            "Bengal is a powerful static site generator. "
+            "Key Features: Fast Builds, Asset Optimization, **SEO** friendly."
+        )
+        result = compute_excerpt(content)
+        from bengal.core.utils.text import strip_html
+
+        plain = strip_html(result)
+        assert not plain.endswith("**"), "Should not end with orphaned **"
+
+    def test_headers_in_content(self):
+        """Headers in content are rendered and available for card excerpt."""
+        content = (
+            "Intro paragraph here.\n\n"
+            "## Key Features\n\n"
+            "### Fast Builds\n"
+            "- Parallel processing\n"
+            "- Asset Optimization\n"
+            "**SEO** friendly."
+        )
+        result = compute_excerpt(content)
+        assert "Key Features" in result or "Intro" in result
